requirements.txt
feedparser>=6.0
google-genai>=1.0
requests>=2.31
python-dotenv>=1.0
